1.1

+++ sources/source.cpp
@@ -25,7 +25,7 @@ std::map<int, int> Cache::number_of_experiments() {
 
 void Cache::straight_pass() {
     int number_of_experimets = _experiments.size();
-    for (unsigned long i = 1; i <= number_of_experimets; ++i) {
+    for (size_t i = 1; i <= number_of_experimets; ++i) {
         set_array(i);
         int size = _experiments[i] / sizeof(int);
         clock_t start = clock();
@@ -42,7 +42,7 @@ void Cache::straight_pass() {
 
 void Cache::back_pass() {
     int number_of_experimets = _experiments.size();
-    for (unsigned long i = 1; i <= number_of_experimets; ++i) {
+    for (size_t i = 1; i <= number_of_experimets; ++i) {
         set_array(i);
         int size = _experiments[i] / sizeof(int);
         clock_t start = clock();
@@ -59,7 +59,7 @@ void Cache::back_pass() {
 
 void Cache::random_pass() {
     int number_of_experimets = _experiments.size();
-    for (unsigned long i = 1; i <= number_of_experimets; ++i) {
+    for (size_t i = 1; i <= number_of_experimets; ++i) {
         set_array(i);
         int size = _experiments[i] / sizeof(int);
         std::vector <size_t> myrand(size);
@@ -80,7 +80,7 @@ void Cache::random_pass() {
 void Cache::set_array(int i) {
     int size = _experiments[i] / sizeof(int);
     _array = new int[size];
-    for (unsigned long k = 0; k < size; ++k) {
+    for (size_t k = 0; k < size; ++k) {
         _array[k] = 0;
     }
 }
@@ -88,7 +88,7 @@ void Cache::set_array(int i) {
 
 void Cache::set_straight(int size) {
     unsigned now = time(0);
-    for (unsigned long i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         _array[i] = rand_r(&now);
     }
 }
@@ -96,14 +96,14 @@ void Cache::set_straight(int size) {
 
 void Cache::set_back(int size) {
     unsigned now = time(0);
-    for (unsigned long i = size - 1; i >= 0; --i) {
+    for (size_t i = size - 1; i >= 0; --i) {
         _array[i] = rand_r(&now);
     }
 }
 
 void Cache::set_random(int size, std::vector <size_t> &myrand) {
     unsigned now = time(0);
-    for (unsigned long i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i) {
         _array[myrand[i]] = rand_r(&now);
     }
 }
